add vacancy management backend, contraints to edit and get

+++ backend/jobsearch/src/main/kotlin/com/jobsearch/service/VacancyService.kt
@@ -1,7 +1,9 @@
 package com.jobsearch.service
 
-import com.jobsearch.dto.VacancyDto
+import com.jobsearch.dto.VacancyRequestDTO
+import com.jobsearch.dto.VacancyResponseDTO
 import com.jobsearch.entity.Vacancy
+import com.jobsearch.exception.ForbiddenException
 import com.jobsearch.exception.NotFoundException
 import com.jobsearch.repository.VacancyRepository
 import jakarta.transaction.Transactional
@@ -13,40 +15,50 @@ class VacancyService(
     val jobFamilyService: JobFamilyService,
     val userService: UserService
 ) {
-    fun createVacancy(vacancyDto: VacancyDto): VacancyDto {
-        val selectedJobFamily = jobFamilyService.findByJobFamilyId(vacancyDto.jobFamilyId!!)
+    fun retrieveVacancy(vacancyId: Int): VacancyResponseDTO {
+        val vacancy = vacancyRepository.findById(vacancyId)
+            .orElseThrow { NotFoundException("No vacancy found with id $vacancyId") }
+        return mapToVacancyResponseDto(vacancy)
+    }
+
+    fun retrieveAllVacancy(): List<VacancyResponseDTO> {
+        val persons = vacancyRepository.findAll()
+        return persons.map {
+            mapToVacancyResponseDto(it)
+        }
+    }
+
+    fun retrieveVacancyByManager(): List<VacancyResponseDTO> {
+        val manager = userService.retrieveAuthenticatedUser()
+        return vacancyRepository.findByManager(manager).map {
+            mapToVacancyResponseDto(it)
+        }
+    }
+
+    @Transactional
+    fun createVacancy(vacancyDto: VacancyRequestDTO): VacancyResponseDTO {
+        val selectedJobFamily = jobFamilyService.findByJobFamilyId(vacancyDto.jobFamilyId)
             .orElseThrow { NotFoundException("No vacancy found with id ${vacancyDto.jobFamilyId}") }
-
         val managerUser = userService.retrieveAuthenticatedUser()
-
         val vacancyEntity = vacancyDto.let {
             Vacancy(it.id, it.name, it.companyName, it.salaryExpectation, it.yearsOfExperience, it.description, selectedJobFamily, managerUser)
         }
+
         val newVacancy = vacancyRepository.save(vacancyEntity)
 
-        return mapToVacancyDto(newVacancy)
+        return mapToVacancyResponseDto(newVacancy)
     }
 
-    fun retrieveVacancy(vacancyId: Int): VacancyDto {
+    @Transactional
+    fun updateVacancy(vacancyId: Int, vacancyDto: VacancyRequestDTO): VacancyResponseDTO {
+
         val vacancy = vacancyRepository.findById(vacancyId)
             .orElseThrow { NotFoundException("No vacancy found with id $vacancyId") }
+        val manager = userService.retrieveAuthenticatedUser()
 
-        return mapToVacancyDto(vacancy)
-    }
+        if (vacancy.manager != manager ) throw ForbiddenException("You are not allowed to edit this vacancy.")
 
-    fun retrieveAllVacancy(): List<VacancyDto> {
-        val persons = vacancyRepository.findAll()
-
-        return persons.map {
-            mapToVacancyDto(it)
-        }
-    }
-
-    fun updateVacancy(vacancyId: Int, vacancyDto: VacancyDto): VacancyDto {
-        val vacancy = vacancyRepository.findById(vacancyId)
-            .orElseThrow { NotFoundException("No vacancy found with id $vacancyId") }
-
-        val selectedJobFamily = jobFamilyService.findByJobFamilyId(vacancyDto.jobFamilyId!!)
+        val selectedJobFamily = jobFamilyService.findByJobFamilyId(vacancyDto.jobFamilyId)
             .orElseThrow { NotFoundException("No vacancy found with id ${vacancyDto.jobFamilyId}") }
 
         vacancy.name = vacancyDto.name
@@ -56,35 +68,39 @@ class VacancyService(
         vacancy.jobFamily = selectedJobFamily
 
         val updatedVacancy = vacancyRepository.save(vacancy)
-        return mapToVacancyDto(updatedVacancy)
+
+        return mapToVacancyResponseDto(updatedVacancy)
     }
 
+    @Transactional
     fun deleteVacancy(vacancyId: Int): String {
         val vacancy = vacancyRepository.findById(vacancyId)
             .orElseThrow { NotFoundException("No vacancy found with id $vacancyId") }
+        val manager = userService.retrieveAuthenticatedUser()
+        if (vacancy.manager != manager ) throw ForbiddenException("You are not allowed to erase this vacancy.")
 
         vacancyRepository.delete(vacancy)
         return "Vacancy deleted successfully"
     }
 
     @Transactional
-    fun findVacanciesByFilter(salary: Int?, jobFamilyId: Int?, yearsOfExperience: Int?): List<VacancyDto> {
+    fun findVacanciesByFilter(salary: Int?, jobFamilyId: Int?, yearsOfExperience: Int?): List<VacancyResponseDTO> {
         val vacancies = vacancyRepository.findVacanciesByFilters(salary, jobFamilyId, yearsOfExperience)
         return vacancies.map {
-            mapToVacancyDto(it)
+            mapToVacancyResponseDto(it)
         }
     }
 
-    fun mapToVacancyDto(vacancy: Vacancy): VacancyDto {
+    fun mapToVacancyResponseDto(vacancy: Vacancy): VacancyResponseDTO {
         return vacancy.let {
-            VacancyDto(
-                it.id,
+            VacancyResponseDTO(
+                it.id!!,
                 it.name,
                 it.companyName,
                 it.salaryExpectation,
                 it.yearsOfExperience,
                 it.description,
-                it.jobFamily.id,
+                it.jobFamily.id!!,
                 it.jobFamily.name,
                 it.manager.id
             )
